Move Entity and EntityReference tests to Node (#4112)
