+++ test/test-component/component.mjs
@@ -1,0 +1,5 @@
+import { Component } from '../../src/framework/components/component.js';
+
+class DummyComponent extends Component {}
+
+export { DummyComponent };
