+++ test/test-component/data.mjs
@@ -1,0 +1,7 @@
+class DummyComponentData {
+    constructor() {
+        this.enabled = true;
+    }
+}
+
+export { DummyComponentData };
