--- tests/helpers/dummy/component.js
@@ -1,8 +1,0 @@
-Object.assign(pc, (function () {
-    var DummyComponent = function DummyComponent() {};
-    DummyComponent = pc.inherits(DummyComponent, pc.Component);
-
-    return {
-        DummyComponent: DummyComponent
-    };
-}()));
