--- tests/helpers/dummy/data.js
@@ -1,7 +1,0 @@
-Object.assign(pc, (function () {
-    var DummyComponentData = function DummyComponentData() {};
-
-    return {
-        DummyComponentData: DummyComponentData
-    };
-}()));
